# perf: use fastcache for zktrie node