--- a/trie/zk_trie_database.go
+++ b/trie/zk_trie_database.go
@@ -42,9 +42,23 @@ func (l *ZktrieDatabase) Get(key []byte) ([]byte, error) {
 	if ok {
 		return value, nil
 	}
+
+	if l.db.cleans != nil {
+		if enc := l.db.cleans.Get(nil, concatKey); enc != nil {
+			memcacheCleanHitMeter.Mark(1)
+			memcacheCleanReadMeter.Mark(int64(len(enc)))
+			return enc, nil
+		}
+	}
+
 	v, err := l.db.diskdb.Get(concatKey)
 	if err == leveldb.ErrNotFound {
 		return nil, ErrNotFound
+	}
+	if l.db.cleans != nil {
+		l.db.cleans.Set(concatKey[:], v)
+		memcacheCleanMissMeter.Mark(1)
+		memcacheCleanWriteMeter.Mark(int64(len(v)))
 	}
 	return v, err
 }