@noel2004 2022-10-31
It seems the cache reduce average op time from 70k to 50k ns in get op. But can anyone explain why the total time for benchmarking increased from 56s to 87s?