@colinlyguo 2022-11-01
> 

The reason seems like the total op in two test cases are different (with before 315362 and after 467114).

BTW, I also run the test cases, with: