@colinlyguo 2022-11-02
> > LGTM. BTW, why in the Get test, the parameters of the levelDB cache and fastcache are 256?
> > https://github.com/scroll-tech/go-ethereum/blob/3144879f65b6bf3c30d27ec2d5c3d32a2932b060/trie/zk_trie_test.go#L162
> > 
> > https://github.com/scroll-tech/go-ethereum/blob/3144879f65b6bf3c30d27ec2d5c3d32a2932b060/trie/zk_trie_test.go#L166
> 
> * The levelDB cache is 256 because I just did a copy from core/state/snapshot/disklayer_test.go:TestDiskSeek. Should I make it bigger?
>   https://github.com/scroll-tech/go-ethereum/blob/3144879f65b6bf3c30d27ec2d5c3d32a2932b060/core/state/snapshot/disklayer_test.go#L528
> * The fastcache is 256 because I use a default value from core/blockchain.go:defaultCacheConfig.TrieCleanLimit.
>   https://github.com/scroll-tech/go-ethereum/blob/3144879f65b6bf3c30d27ec2d5c3d32a2932b060/core/blockchain.go#L143

I see, and thanks.