@shu-unifra 2022-11-02
> It seems the cache reduce average op time from 70k to 50k ns in get op. But can anyone explain why the total time for benchmarking increased from 56s to 87s?

It's a good question. Even [colinlyguo](https://github.com/colinlyguo) run the test in his computer and it seems the total time is close. But on my computer, the result is always 56s and 87s. I will run more tests and figure it out.