@colinlyguo 2022-11-02
> > It seems the cache reduce average op time from 70k to 50k ns in get op. But can anyone explain why the total time for benchmarking increased from 56s to 87s?
> 
> I did more tests. It shows that the 2 cases run different round on my computer, the b.N increase every round. There are 4 rounds when I run "before" and 5 rounds when I run "after". Before "b.ResetTimer()" it create a new zktrie and puts 10000 leafs into the trie which takes a lot of time(8s on my computer).
> 
> * 87-56=31s, close to 24s+8s;
> * 24s is "actual -benchtime", i.e. (492546*48699)ns
> * 8s is the time we build a 10k-leafs zktrie.

This makes sense I think, when I run the tests, the b.N increases:
**Before:**
b.N: 100
b.N: 10000
b.N: 652005
**After:**
b.N: 100
b.N: 10000
b.N: 779097

And the doc: https://pkg.go.dev/testing